+++ Normaize.Core/Interfaces/IChartGenerationService.cs
@@ -18,7 +18,7 @@ public interface IChartGenerationService
     /// <param name="configuration">Optional chart configuration</param>
     /// <param name="context">Operation context for logging</param>
     /// <returns>Chart data DTO</returns>
-    ChartDataDto GenerateChartData(DataSet dataSet, List<Dictionary<string, object>> data, ChartType chartType, ChartConfigurationDto? configuration, IOperationContext context);
+    ChartDataDto GenerateChartData(DataSet dataSet, List<Dictionary<string, object?>> data, ChartType chartType, ChartConfigurationDto? configuration, IOperationContext context);
 
     /// <summary>
     /// Generates comparison chart data for two datasets.
@@ -31,7 +31,7 @@ public interface IChartGenerationService
     /// <param name="configuration">Optional chart configuration</param>
     /// <param name="context">Operation context for logging</param>
     /// <returns>Comparison chart data DTO</returns>
-    ComparisonChartDto GenerateComparisonChartData(DataSet dataSet1, DataSet dataSet2, List<Dictionary<string, object>> data1, List<Dictionary<string, object>> data2, ChartType chartType, ChartConfigurationDto? configuration, IOperationContext context);
+    ComparisonChartDto GenerateComparisonChartData(DataSet dataSet1, DataSet dataSet2, List<Dictionary<string, object?>> data1, List<Dictionary<string, object?>> data2, ChartType chartType, ChartConfigurationDto? configuration, IOperationContext context);
 
     /// <summary>
     /// Validates chart configuration for a specific chart type.
