+++ Normaize.Tests/Services/FileUploadServiceTests.cs
@@ -16,7 +16,12 @@ namespace Normaize.Tests.Services;
 
 public class FileUploadServiceTests
 {
-    private readonly Mock<IStorageService> _mockStorageService;
+    private readonly Mock<IFileUploadServices> _mockFileUploadServices;
+    private readonly Mock<IFileValidationService> _mockValidationService;
+    private readonly Mock<IFileProcessingService> _mockProcessingService;
+    private readonly Mock<IFileConfigurationService> _mockConfigurationService;
+    private readonly Mock<IFileUtilityService> _mockUtilityService;
+    private readonly Mock<IFileStorageService> _mockFileStorageService;
     private readonly Mock<IDataProcessingInfrastructure> _mockInfrastructure;
     private readonly FileUploadService _service;
     private readonly FileUploadConfiguration _fileUploadConfig;
@@ -24,15 +29,12 @@ public class FileUploadServiceTests
 
     public FileUploadServiceTests()
     {
-        _mockStorageService = new Mock<IStorageService>();
-        _mockInfrastructure = new Mock<IDataProcessingInfrastructure>();
-
         // Setup configuration objects
         _fileUploadConfig = new FileUploadConfiguration
         {
             MaxFileSize = 10485760, // 10MB
             AllowedExtensions = new[] { ".csv", ".json", ".xlsx", ".xls", ".xml", ".parquet", ".txt" },
-            MaxPreviewRows = 100, // Must be at least 100 (matches new default)
+            MaxPreviewRows = 100,
             MaxConcurrentUploads = 5,
             EnableCompression = true,
             BlockedExtensions = new[] { ".exe", ".bat", ".cmd", ".ps1", ".sh", ".dll", ".so", ".dylib" }
@@ -42,25 +44,37 @@ public class FileUploadServiceTests
         {
             MaxRowsPerDataset = 10000,
             MaxColumnsPerDataset = 100,
-            MaxPreviewRows = 10, // Must be between 1-100
+            MaxPreviewRows = 10,
             EnableDataValidation = true,
             EnableSchemaInference = true,
             MaxProcessingTimeSeconds = 30
         };
 
-        var mockFileUploadOptions = new Mock<IOptions<FileUploadConfiguration>>();
-        mockFileUploadOptions.Setup(x => x.Value).Returns(_fileUploadConfig);
-
-        var mockDataProcessingOptions = new Mock<IOptions<DataProcessingConfiguration>>();
-        mockDataProcessingOptions.Setup(x => x.Value).Returns(_dataProcessingConfig);
+        // Create mock sub-services
+        _mockValidationService = new Mock<IFileValidationService>();
+        _mockProcessingService = new Mock<IFileProcessingService>();
+        _mockConfigurationService = new Mock<IFileConfigurationService>();
+        _mockUtilityService = new Mock<IFileUtilityService>();
+        _mockFileStorageService = new Mock<IFileStorageService>();
+        _mockInfrastructure = new Mock<IDataProcessingInfrastructure>();
 
         // Setup infrastructure mocks
         SetupInfrastructureMocks();
 
+        // Setup configuration service to handle constructor calls
+        _mockConfigurationService.Setup(x => x.ValidateConfiguration()).Verifiable();
+        _mockConfigurationService.Setup(x => x.LogConfiguration()).Verifiable();
+
+        // Setup mock composite service
+        _mockFileUploadServices = new Mock<IFileUploadServices>();
+        _mockFileUploadServices.Setup(x => x.Validation).Returns(_mockValidationService.Object);
+        _mockFileUploadServices.Setup(x => x.Processing).Returns(_mockProcessingService.Object);
+        _mockFileUploadServices.Setup(x => x.Configuration).Returns(_mockConfigurationService.Object);
+        _mockFileUploadServices.Setup(x => x.Utility).Returns(_mockUtilityService.Object);
+        _mockFileUploadServices.Setup(x => x.Storage).Returns(_mockFileStorageService.Object);
+
         _service = new FileUploadService(
-            mockFileUploadOptions.Object,
-            mockDataProcessingOptions.Object,
-            _mockStorageService.Object,
+            _mockFileUploadServices.Object,
             _mockInfrastructure.Object);
     }
 
@@ -72,34 +86,45 @@ public class FileUploadServiceTests
         // Arrange
         var fileRequest = CreateValidFileUploadRequest();
         var expectedPath = "uploads/test.csv";
-        _mockStorageService.Setup(x => x.SaveFileAsync(fileRequest)).ReturnsAsync(expectedPath);
+
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(true);
+        _mockFileStorageService.Setup(x => x.SaveFileAsync(fileRequest)).ReturnsAsync(expectedPath);
 
         // Act
         var result = await _service.SaveFileAsync(fileRequest);
 
         // Assert
         Assert.Equal(expectedPath, result);
-        _mockStorageService.Verify(x => x.SaveFileAsync(fileRequest), Times.Once);
+        _mockValidationService.Verify(x => x.ValidateFileAsync(fileRequest), Times.Once);
+        _mockFileStorageService.Verify(x => x.SaveFileAsync(fileRequest), Times.Once);
+    }
+
+    [Fact]
+    public async Task SaveFileAsync_ValidationFails_ThrowsException()
+    {
+        // Arrange
+        var fileRequest = CreateValidFileUploadRequest();
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(false);
+
+        // Act & Assert
+        await Assert.ThrowsAsync<FileValidationException>(() =>
+            _service.SaveFileAsync(fileRequest));
     }
 
     [Fact]
     public async Task SaveFileAsync_StorageServiceThrowsException_PropagatesException()
     {
         // Arrange
-        var fileRequest = new FileUploadRequest
-        {
-            FileName = "test.csv",
-            FileSize = 1024,
-            FileStream = new MemoryStream(Encoding.UTF8.GetBytes("test"))
-        };
+        var fileRequest = CreateValidFileUploadRequest();
         var expectedException = new InvalidOperationException("Storage error");
-        _mockStorageService.Setup(x => x.SaveFileAsync(fileRequest)).ThrowsAsync(expectedException);
+
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(true);
+        _mockFileStorageService.Setup(x => x.SaveFileAsync(fileRequest)).ThrowsAsync(expectedException);
 
         // Act & Assert
         var exception = await Assert.ThrowsAsync<FileUploadException>(() =>
             _service.SaveFileAsync(fileRequest));
-        Assert.Contains("Failed to save file test.csv", exception.Message);
-        Assert.Equal(expectedException, exception.InnerException);
+        Assert.Contains("Failed to save file", exception.Message);
     }
 
     #endregion
@@ -107,72 +132,33 @@ public class FileUploadServiceTests
     #region ValidateFileAsync Tests
 
     [Fact]
-    public async Task ValidateFileAsync_FileSizeExceedsLimit_ReturnsFalse()
-    {
-        // Arrange
-        var fileRequest = new FileUploadRequest
-        {
-            FileName = "large.csv",
-            FileSize = 20000000 // 20MB, exceeds 10MB limit
-        };
-
-        // Act
-        var result = await _service.ValidateFileAsync(fileRequest);
-
-        // Assert
-        Assert.False(result);
-    }
-
-    [Fact]
-    public async Task ValidateFileAsync_BlockedExtension_ReturnsFalse()
-    {
-        // Arrange
-        var fileRequest = new FileUploadRequest
-        {
-            FileName = "malicious.exe",
-            FileSize = 1024
-        };
-
-        // Act
-        var result = await _service.ValidateFileAsync(fileRequest);
-
-        // Assert
-        Assert.False(result);
-    }
-
-    [Fact]
     public async Task ValidateFileAsync_ValidFile_ReturnsTrue()
     {
         // Arrange
-        var fileRequest = new FileUploadRequest
-        {
-            FileName = "test.csv",
-            FileSize = 1024
-        };
+        var fileRequest = CreateValidFileUploadRequest();
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(true);
 
         // Act
         var result = await _service.ValidateFileAsync(fileRequest);
 
         // Assert
         Assert.True(result);
+        _mockValidationService.Verify(x => x.ValidateFileAsync(fileRequest), Times.Once);
     }
 
-    [Theory]
-    [InlineData(".pdf")]
-    [InlineData(".doc")]
-    [InlineData(".exe")]
-    [InlineData(".zip")]
-    public async Task ValidateFileAsync_InvalidExtension_ReturnsFalse(string invalidExtension)
+    [Fact]
+    public async Task ValidateFileAsync_InvalidFile_ReturnsFalse()
     {
         // Arrange
         var fileRequest = CreateValidFileUploadRequest();
-        fileRequest.FileName = $"test{invalidExtension}";
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(false);
 
         // Act
         var result = await _service.ValidateFileAsync(fileRequest);
 
         // Assert
         Assert.False(result);
+        _mockValidationService.Verify(x => x.ValidateFileAsync(fileRequest), Times.Once);
     }
 
     [Theory]
@@ -188,12 +174,34 @@ public class FileUploadServiceTests
         // Arrange
         var fileRequest = CreateValidFileUploadRequest();
         fileRequest.FileName = $"test{validExtension}";
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(true);
 
         // Act
         var result = await _service.ValidateFileAsync(fileRequest);
 
         // Assert
         Assert.True(result);
+        _mockValidationService.Verify(x => x.ValidateFileAsync(fileRequest), Times.Once);
+    }
+
+    [Theory]
+    [InlineData(".pdf")]
+    [InlineData(".doc")]
+    [InlineData(".exe")]
+    [InlineData(".zip")]
+    public async Task ValidateFileAsync_InvalidExtension_ReturnsFalse(string invalidExtension)
+    {
+        // Arrange
+        var fileRequest = CreateValidFileUploadRequest();
+        fileRequest.FileName = $"test{invalidExtension}";
+        _mockValidationService.Setup(x => x.ValidateFileAsync(fileRequest)).ReturnsAsync(false);
+
+        // Act
+        var result = await _service.ValidateFileAsync(fileRequest);
+
+        // Assert
+        Assert.False(result);
+        _mockValidationService.Verify(x => x.ValidateFileAsync(fileRequest), Times.Once);
     }
 
     #endregion
@@ -201,12 +209,37 @@ public class FileUploadServiceTests
     #region ProcessFileAsync Tests
 
     [Fact]
+    public async Task ProcessFileAsync_ValidFile_ReturnsDataSet()
+    {
+        // Arrange
+        var filePath = "test.csv";
+        var fileType = ".csv";
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.csv",
+            FileType = FileType.CSV,
+            RowCount = 10,
+            ColumnCount = 5
+        };
+
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
+
+        // Act
+        var result = await _service.ProcessFileAsync(filePath, fileType);
+
+        // Assert
+        Assert.Equal(expectedDataSet, result);
+        _mockProcessingService.Verify(x => x.ProcessFileAsync(filePath, fileType), Times.Once);
+    }
+
+    [Fact]
     public async Task ProcessFileAsync_FileNotFound_ThrowsFileNotFoundException()
     {
         // Arrange
         var filePath = "nonexistent.csv";
         var fileType = ".csv";
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(false);
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType))
+            .ThrowsAsync(new FileNotFoundException("File not found"));
 
         // Act & Assert
         var exception = await Assert.ThrowsAsync<FileNotFoundException>(() =>
@@ -220,7 +253,8 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.unsupported";
         var fileType = ".unsupported";
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType))
+            .ThrowsAsync(new UnsupportedFileTypeException("File type .unsupported is not supported"));
 
         // Act & Assert
         var exception = await Assert.ThrowsAsync<UnsupportedFileTypeException>(() =>
@@ -229,26 +263,19 @@ public class FileUploadServiceTests
     }
 
     [Fact]
-    public async Task ProcessFileAsync_ProcessingError_SetsErrorProperties()
+    public async Task ProcessFileAsync_ProcessingError_ThrowsException()
     {
         // Arrange
         var filePath = "test.csv";
         var fileType = ".csv";
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ThrowsAsync(new IOException("Read error"));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType))
+            .ThrowsAsync(new FileProcessingException("Processing error"));
 
-        // Act
-        var result = await _service.ProcessFileAsync(filePath, fileType);
-
-        // Assert
-        Assert.False(result.IsProcessed);
-        Assert.Contains("Read error", result.ProcessingErrors);
-        Assert.NotNull(result.ProcessingErrors);
+        // Act & Assert
+        var exception = await Assert.ThrowsAsync<FileProcessingException>(() =>
+            _service.ProcessFileAsync(filePath, fileType));
+        Assert.Contains("Processing error", exception.Message);
     }
-
-    #endregion
-
-    #region CSV Processing Tests
 
     [Fact]
     public async Task ProcessFileAsync_CsvFile_ProcessesCorrectly()
@@ -256,133 +283,49 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.csv";
         var fileType = ".csv";
-        var csvContent = "Name,Age,City\nJohn,30,New York\nJane,25,Boston";
-        var csvBytes = Encoding.UTF8.GetBytes(csvContent);
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.csv",
+            FileType = FileType.CSV,
+            RowCount = 5,
+            ColumnCount = 3
+        };
 
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(csvBytes));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
 
         // Act
         var result = await _service.ProcessFileAsync(filePath, fileType);
 
         // Assert
-        if (!result.IsProcessed)
-        {
-            // Output the error for debugging
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
-        Assert.Equal("test.csv", result.FileName);
-        Assert.Equal(filePath, result.FilePath);
+        Assert.Equal(expectedDataSet, result);
         Assert.Equal(FileType.CSV, result.FileType);
-        Assert.Equal(2, result.RowCount);
+        Assert.Equal(5, result.RowCount);
         Assert.Equal(3, result.ColumnCount);
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
-        Assert.NotNull(result.DataHash);
     }
 
     [Fact]
-    public async Task ProcessFileAsync_CsvFileWithManyRows_UsesSeparateTable()
-    {
-        // Arrange
-        var filePath = "large.csv";
-        var fileType = ".csv";
-        var csvContent = "Name,Age\n" + string.Join("\n", Enumerable.Range(1, 15000).Select(i => $"User{i},{i}"));
-        var csvBytes = Encoding.UTF8.GetBytes(csvContent);
-
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(csvBytes));
-
-        // Act
-        var result = await _service.ProcessFileAsync(filePath, fileType);
-
-        // Assert
-        if (!result.UseSeparateTable)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"RowCount: {result.RowCount}, MaxRows: {10000}");
-            System.Console.WriteLine($"FileSize: {result.FileSize}, MaxFileSize: {10485760}");
-        }
-        Assert.True(result.UseSeparateTable);
-        Assert.Equal(10000, result.RowCount); // Limited by maxRows
-        Assert.Null(result.ProcessedData); // Should not store inline
-        Assert.NotNull(result.PreviewData); // Should still have preview
-    }
-
-    #endregion
-
-    #region JSON Processing Tests
-
-    [Fact]
-    public async Task ProcessFileAsync_JsonArray_ProcessesCorrectly()
+    public async Task ProcessFileAsync_JsonFile_ProcessesCorrectly()
     {
         // Arrange
         var filePath = "test.json";
         var fileType = ".json";
-        var jsonContent = @"[
-            {""name"": ""John"", ""age"": 30, ""city"": ""New York""},
-            {""name"": ""Jane"", ""age"": 25, ""city"": ""Boston""}
-        ]";
-        var jsonBytes = Encoding.UTF8.GetBytes(jsonContent);
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.json",
+            FileType = FileType.JSON,
+            RowCount = 10,
+            ColumnCount = 4
+        };
 
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(jsonBytes));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
 
         // Act
         var result = await _service.ProcessFileAsync(filePath, fileType);
 
         // Assert
-        if (!result.IsProcessed)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
+        Assert.Equal(expectedDataSet, result);
         Assert.Equal(FileType.JSON, result.FileType);
-        Assert.Equal(2, result.RowCount);
-        Assert.Equal(3, result.ColumnCount);
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
     }
-
-    [Fact]
-    public async Task ProcessFileAsync_JsonObject_ProcessesCorrectly()
-    {
-        // Arrange
-        var filePath = "test.json";
-        var fileType = ".json";
-        var jsonContent = @"{""name"": ""John"", ""age"": 30, ""city"": ""New York""}";
-        var jsonBytes = Encoding.UTF8.GetBytes(jsonContent);
-
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(jsonBytes));
-
-        // Act
-        var result = await _service.ProcessFileAsync(filePath, fileType);
-
-        // Assert
-        if (!result.IsProcessed)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
-        Assert.Equal(FileType.JSON, result.FileType);
-        Assert.Equal(1, result.RowCount);
-        Assert.Equal(3, result.ColumnCount);
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
-    }
-
-    #endregion
-
-    #region Excel Processing Tests
 
     [Fact]
     public async Task ProcessFileAsync_ExcelFile_ProcessesCorrectly()
@@ -390,32 +333,23 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.xlsx";
         var fileType = ".xlsx";
-        var excelBytes = CreateSampleExcelFile();
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.xlsx",
+            FileType = FileType.Excel,
+            RowCount = 15,
+            ColumnCount = 6
+        };
 
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(excelBytes));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
 
         // Act
         var result = await _service.ProcessFileAsync(filePath, fileType);
 
         // Assert
-        if (!result.IsProcessed)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
+        Assert.Equal(expectedDataSet, result);
         Assert.Equal(FileType.Excel, result.FileType);
-        Assert.Equal(2, result.RowCount);
-        Assert.Equal(3, result.ColumnCount);
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
     }
-
-    #endregion
-
-    #region XML Processing Tests
 
     [Fact]
     public async Task ProcessFileAsync_XmlFile_ProcessesCorrectly()
@@ -423,45 +357,23 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.xml";
         var fileType = ".xml";
-        var xmlContent = @"<?xml version=""1.0"" encoding=""UTF-8""?>
-<root>
-    <item>
-        <name>John</name>
-        <age>30</age>
-        <city>New York</city>
-    </item>
-    <item>
-        <name>Jane</name>
-        <age>25</age>
-        <city>Boston</city>
-    </item>
-</root>";
-        var xmlBytes = Encoding.UTF8.GetBytes(xmlContent);
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.xml",
+            FileType = FileType.XML,
+            RowCount = 8,
+            ColumnCount = 5
+        };
 
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(xmlBytes));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
 
         // Act
         var result = await _service.ProcessFileAsync(filePath, fileType);
 
         // Assert
-        if (!result.IsProcessed)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
+        Assert.Equal(expectedDataSet, result);
         Assert.Equal(FileType.XML, result.FileType);
-        Assert.Equal(2, result.RowCount);
-        Assert.Equal(3, result.ColumnCount);
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
     }
-
-    #endregion
-
-    #region TXT Processing Tests
 
     [Fact]
     public async Task ProcessFileAsync_TextFile_ProcessesCorrectly()
@@ -469,28 +381,22 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.txt";
         var fileType = ".txt";
-        var textContent = "Line 1\nLine 2\nLine 3";
-        var textBytes = Encoding.UTF8.GetBytes(textContent);
+        var expectedDataSet = new DataSet
+        {
+            FileName = "test.txt",
+            FileType = FileType.TXT,
+            RowCount = 20,
+            ColumnCount = 1
+        };
 
-        _mockStorageService.Setup(x => x.FileExistsAsync(filePath)).ReturnsAsync(true);
-        _mockStorageService.Setup(x => x.GetFileAsync(filePath)).ReturnsAsync(() => new MemoryStream(textBytes));
+        _mockProcessingService.Setup(x => x.ProcessFileAsync(filePath, fileType)).ReturnsAsync(expectedDataSet);
 
         // Act
         var result = await _service.ProcessFileAsync(filePath, fileType);
 
         // Assert
-        if (!result.IsProcessed)
-        {
-            System.Diagnostics.Debug.WriteLine($"Processing error: {result.ProcessingErrors}");
-            System.Console.WriteLine($"Processing error: {result.ProcessingErrors}");
-        }
-        Assert.True(result.IsProcessed);
+        Assert.Equal(expectedDataSet, result);
         Assert.Equal(FileType.TXT, result.FileType);
-        Assert.Equal(3, result.RowCount);
-        Assert.Equal(2, result.ColumnCount); // LineNumber and Content columns
-        Assert.NotNull(result.Schema);
-        Assert.NotNull(result.PreviewData);
-        Assert.NotNull(result.ProcessedData);
     }
 
     #endregion
@@ -502,13 +408,13 @@ public class FileUploadServiceTests
     {
         // Arrange
         var filePath = "test.csv";
-        _mockStorageService.Setup(x => x.DeleteFileAsync(filePath)).Returns(Task.CompletedTask);
+        _mockFileStorageService.Setup(x => x.DeleteFileAsync(filePath)).Returns(Task.CompletedTask);
 
         // Act
         await _service.DeleteFileAsync(filePath);
 
         // Assert
-        _mockStorageService.Verify(x => x.DeleteFileAsync(filePath), Times.Once);
+        _mockFileStorageService.Verify(x => x.DeleteFileAsync(filePath), Times.Once);
     }
 
     [Fact]
@@ -517,14 +423,11 @@ public class FileUploadServiceTests
         // Arrange
         var filePath = "test.csv";
         var expectedException = new InvalidOperationException("Delete error");
-        _mockStorageService.Setup(x => x.DeleteFileAsync(filePath)).ThrowsAsync(expectedException);
+        _mockFileStorageService.Setup(x => x.DeleteFileAsync(filePath)).ThrowsAsync(expectedException);
 
-        // Act
+        // Act & Assert - should not throw, just log error
         await _service.DeleteFileAsync(filePath);
-
-        // Assert
-        // The service catches and logs exceptions, doesn't re-throw them
-        _mockStorageService.Verify(x => x.DeleteFileAsync(filePath), Times.Once);
+        _mockFileStorageService.Verify(x => x.DeleteFileAsync(filePath), Times.Once);
     }
 
     #endregion
@@ -533,40 +436,22 @@ public class FileUploadServiceTests
 
     private void SetupInfrastructureMocks()
     {
-        // Setup structured logging mock
         var mockStructuredLogging = new Mock<IStructuredLoggingService>();
-        _mockInfrastructure.Setup(x => x.StructuredLogging).Returns(mockStructuredLogging.Object);
-
-        // Setup chaos engineering mock
         var mockChaosEngineering = new Mock<IChaosEngineeringService>();
+        var mockOperationContext = new Mock<IOperationContext>();
+
+        _mockInfrastructure.Setup(x => x.StructuredLogging).Returns(mockStructuredLogging.Object);
         _mockInfrastructure.Setup(x => x.ChaosEngineering).Returns(mockChaosEngineering.Object);
 
-        // Setup default timeout values
-        _mockInfrastructure.Setup(x => x.DefaultTimeout).Returns(TimeSpan.FromSeconds(30));
-        _mockInfrastructure.Setup(x => x.QuickTimeout).Returns(TimeSpan.FromSeconds(5));
+        mockStructuredLogging.Setup(x => x.CreateContext(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Dictionary<string, object>>()))
+            .Returns(mockOperationContext.Object);
+        mockStructuredLogging.Setup(x => x.LogStep(It.IsAny<IOperationContext>(), It.IsAny<string>(), It.IsAny<Dictionary<string, object>>())).Verifiable();
+        mockStructuredLogging.Setup(x => x.LogSummary(It.IsAny<IOperationContext>(), It.IsAny<bool>(), It.IsAny<string>())).Verifiable();
 
-        // Setup structured logging context creation
-        mockStructuredLogging.Setup(s => s.CreateContext(It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Dictionary<string, object>>()))
-            .Returns<string, string, string, Dictionary<string, object>>((operationName, correlationId, userId, metadata) =>
-            {
-                var mockContext = new Mock<IOperationContext>();
-                mockContext.Setup(c => c.OperationName).Returns(operationName);
-                mockContext.Setup(c => c.CorrelationId).Returns(correlationId);
-                mockContext.Setup(c => c.UserId).Returns(userId);
-                mockContext.Setup(c => c.Metadata).Returns(metadata ?? []);
-                mockContext.Setup(c => c.Steps).Returns([]);
-                mockContext.Setup(c => c.Stopwatch).Returns(System.Diagnostics.Stopwatch.StartNew());
-                return mockContext.Object;
-            });
-
-        // Setup chaos engineering execution
         mockChaosEngineering.Setup(x => x.ExecuteChaosAsync(
-            It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(), It.IsAny<Func<Task>>(), It.IsAny<Dictionary<string, object>>()))
-            .Returns<string, string, string, Func<Task>, Dictionary<string, object>>(async (scenario, correlationId, operationName, operation, metadata) =>
-            {
-                await operation();
-                return true;
-            });
+            It.IsAny<string>(), It.IsAny<string>(), It.IsAny<string>(),
+            It.IsAny<Func<Task>>(), It.IsAny<Dictionary<string, object>>()))
+            .ReturnsAsync(true);
     }
 
     private static FileUploadRequest CreateValidFileUploadRequest()
@@ -574,37 +459,9 @@ public class FileUploadServiceTests
         return new FileUploadRequest
         {
             FileName = "test.csv",
-            ContentType = "text/csv",
             FileSize = 1024,
-            FileStream = new MemoryStream(Encoding.UTF8.GetBytes("test content")),
-            Description = "Test file",
-            Tags = "test,csv",
-            KeepOriginalFile = true,
-            StoreProcessedData = true
+            FileStream = new MemoryStream(Encoding.UTF8.GetBytes("test,data\n1,2\n3,4"))
         };
-    }
-
-    private static byte[] CreateSampleExcelFile()
-    {
-        // Create a simple Excel file for testing
-        using var package = new OfficeOpenXml.ExcelPackage();
-        var worksheet = package.Workbook.Worksheets.Add("Sheet1");
-
-        // Add headers
-        worksheet.Cells[1, 1].Value = "Name";
-        worksheet.Cells[1, 2].Value = "Age";
-        worksheet.Cells[1, 3].Value = "City";
-
-        // Add data
-        worksheet.Cells[2, 1].Value = "John";
-        worksheet.Cells[2, 2].Value = 30;
-        worksheet.Cells[2, 3].Value = "New York";
-
-        worksheet.Cells[3, 1].Value = "Jane";
-        worksheet.Cells[3, 2].Value = 25;
-        worksheet.Cells[3, 3].Value = "Boston";
-
-        return package.GetAsByteArray();
     }
 
     #endregion
